Move kansa/member/ to client/
